pyproject.toml
[tool.poetry]
name = "py-weathercn"
version = "0.1.0"
description = "中国天气python插件"
authors = ["ssfdust <ssfdust@gmail.com>"]
readme = "README.md"
packages = [
    {include = "weathercn"}
]

[tool.poetry.scripts]
weathercn = "weathercn.main:main"

[tool.poetry.dependencies]
python = "^3.7"
requests = "^2.21"
css-parser = "^1.0"
pillow = "^6.0"
regex = "^2019.4"
arrow = "^0.13.1"

[tool.poetry.dev-dependencies]
ipdb = "^0.12.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
